Biased Genes Identification

Function to identify the bias genes based on user-selected threshold of number of standard deviation in relative change in deviance and rank difference.

Usage

biasDetect(list_batch_df, threshold = "both", nSD_dev = NULL, nSD_rank = NULL)

Arguments

list_batch_df

list : The list of data frame(s) generated from featureSelection() function. The length of the data frame list should be at least one.

threshold

A character string specifying the filtering criterion. Must be one of:

"dev": Filters genes based on the deviance threshold only.

"rank": Filters genes based on the rank threshold only.

"both": Filters genes based on either the deviance or rank threshold. Default is "both".

nSD_dev

integer: A numeric vector specifying the number of standard deviation (nSD) for each batch when analyzing the relative change in deviance. The order of values must correspond to the order of batches in list_batch_df. Required if threshold is "dev" or "both". If a single value is provided, it is applied to all batches; otherwise, it must have the same length as list_batch_df.

nSD_rank

vector: A numeric vector specifying the number of standard deviation (nSD) for each batch when analyzing rank differences. The order of values must correspond to the order of batches in list_batch_df. Required if threshold is "rank" or "both". If a single value is provided, it is applied to all batches; otherwise, it must have the same length as list_batch_df.

Value

A named list where each element corresponds to a batch and contains:

"Plot": A diagnostic plot (either deviance, rank, or both).

"Table": A filtered data frame containing outlier genes based on the specified threshold.

Examples

data(spe_sub4)
#> Warning: file ‘spe_sub4.rda’ has magic number 'versi'
#>   Use of save versions prior to 2 is deprecated
#> Error in load(file, envir = tmp_env): bad restore file magic number (file may be corrupted) -- no data loaded
spe_sub4
#> Error: object 'spe_sub4' not found

data(svgs_sub4)
#> Warning: file ‘svgs_sub4.rda’ has magic number 'versi'
#>   Use of save versions prior to 2 is deprecated
#> Error in load(file, envir = tmp_env): bad restore file magic number (file may be corrupted) -- no data loaded
SVGs <- svgs_sub4$gene_id
#> Error: object 'svgs_sub4' not found

list_batch_df <- featureSelect(input = spe_sub4, 
    batch_effects = c("sample_id", "sex"), VGs = SVGs)
#> Error: object 'spe_sub4' not found

biaGenes <- biasDetect(list_batch_df = list_batch_df, threshold = "both", 
    nSD_dev = c(5,4), nSD_rank = c(4,6))
#> Error: object 'list_batch_df' not found