Feature selection

The function computes batch-adjusted deviance values, ranks the genes accordingly, and quantifies batch effects in terms of standard deviations from the mean difference. The list follows the order of batch effects provided in batch_effects.

Usage

featureSelect(input, batch_effects = NULL, VGs = NULL)

Arguments

input

A SpatialExperiment object containing spatial transcriptomics data..

batch_effects

A character vector specifying column names in colData(input) that indicate batch effects. Must match existing column names.

VGs

A character vector specifying the variable genes (VGs) to be analyzed. Only genes present in this vector will be retained for feature selection.

Value

A named list where each element corresponds to a batch effect. Each batch contains a data frame with the following columns:

"gene_id": Gene identifier.

"gene_name": Gene name.

"dev_default": Deviance score without batch correction.

"dev_": Deviance score with batch correction.

"rank_default": Rank of the gene based on deviance without batch correction.

"rank_": Rank of the gene based on deviance with batch correction.

"d_diff": Relative change in deviance between default and batch-corrected models.

"nSD_dev_": number of standard deviation of relative change in deviance for the batch.

"r_diff": Rank difference between default and batch-corrected models.

"nSD_rank_": number of standard deviation of rank difference for the batch.

Examples

data(spe_sub4)
#> Warning: file ‘spe_sub4.rda’ has magic number 'versi'
#>   Use of save versions prior to 2 is deprecated
#> Error in load(file, envir = tmp_env): bad restore file magic number (file may be corrupted) -- no data loaded
spe_sub4
#> Error: object 'spe_sub4' not found

data(svgs_sub4)
#> Warning: file ‘svgs_sub4.rda’ has magic number 'versi'
#>   Use of save versions prior to 2 is deprecated
#> Error in load(file, envir = tmp_env): bad restore file magic number (file may be corrupted) -- no data loaded
SVGs <- svgs_sub4$gene_id
#> Error: object 'svgs_sub4' not found

batch_df <- featureSelect(input = spe_sub4, 
    batch_effects = c("sample_id", "sex"), VGs = SVGs)
#> Error: object 'spe_sub4' not found