Example SpatialExperiment Dataset for Batch Effect Analysis

spe_sub4 is an example SpatialExperiment object used to demonstrate batch effect correction and spatially variable gene (SVG) selection in the BatchSVG package. It takes the four samples ("V11L05-333_B1", "V11L05-333_D1","V11L05-335_D1","V11L05-336_A1") from the full spatial transcriptomics data stored in the humanHippocampus2024 data package.

Usage

data("svgs_sub4")

Format

A SpatialExperiment object with:

Dimensions

2082 genes (rows) × 18,945 spots (columns).

Assays

counts (raw gene expression) and logcounts (log-transformed normalized counts).

Row metadata (rowData)

Includes gene annotations such as source, type, and gene_search.

Column metadata (colData)

Includes 150 variables such as sample_id, in_tissue, and nmf100.

Reduced dimensions

Contains PCA, t-SNE, and UMAP embeddings labeled as 10x_pca, 10x_tsne, and 10x_umap.

Spatial coordinates

Available in spatialCoords with pixel row and column positions.

Image metadata

Includes fields such as sample_id, image_id, and scaleFactor.

Source

Processed spatial transcriptomics data for demonstration purposes.

Examples

data("svgs_sub4")
#> Warning: file ‘svgs_sub4.rda’ has magic number 'versi'
#>   Use of save versions prior to 2 is deprecated
#> Error in load(file, envir = tmp_env): bad restore file magic number (file may be corrupted) -- no data loaded