Biased Genes Identification

Function to identify the bias genes based on user-selected threshold of number of standard deviation in relative change in deviance and rank difference.

Usage

biasDetect(list_batch_df, nSD_dev, nSD_rank)

Arguments

list_batch_df

list : The list of data frame(s) generated from featureSelection() function. The length of the data frame list should be at least one.

nSD_dev

integer: A numeric vector specifying the number of standard deviation (nSD) for each batch when analyzing the relative change in deviance. The order of values must correspond to the order of batches in list_batch_df. If a single value is provided, it is applied to all batches; otherwise, it must have the same length as list_batch_df.

nSD_rank

vector: A numeric vector specifying the number of standard deviation (nSD) for each batch when analyzing rank differences. The order of values must correspond to the order of batches in list_batch_df. If a single value is provided, it is applied to all batches; otherwise, it must have the same length as list_batch_df.

visual

logical: Whether to display the detected bias genes by visualizations of deviance and rank. Default is FALSE. If visual = TRUE, the returned format will be two parallel plots presenting bias genes based on both nSD_dev and nSD_rank.

Value

The named list where each element corresponds to a batch. Each batch contains:

"Plot": A combined ggplot object displaying two scatter plots—one for deviance and one for rank—with an enforced 1:1 aspect ratio.

"Table": A data frame containing identified biased genes, with columns indicating whether they are outliers based on deviance or rank thresholds.

The order of batches in the list follows the order of nSD_dev and nSD_rank provided by the user.

Examples

load(system.file("extdata","list_batch_df.rda",package = "BatchSVG"))
#> Warning: cannot open compressed file '', probable reason 'No such file or directory'
#> Error in readChar(con, 5L, useBytes = TRUE): cannot open the connection

biaGenes <- biasDetect(list_batch_df, nSD_dev = c(5,4), nSD_rank = c(4,6))
#> Error: object 'list_batch_df' not found

biaGenes[["sample_id"]][["Table"]] # see biased genes in table format
#> Error: object 'biaGenes' not found
biaGenes[["sex"]][["Plot"]] # see biased genes in plot format
#> Error: object 'biaGenes' not found